Distribution S4 Classes

A collection of S4 classes that provide a flexible and structured way to work with probability distributions.

Usage

d(distr, x, ...)

p(distr, q, ...)

qn(distr, p, ...)

r(distr, n, ...)

Arguments

distr

an object of class Distribution or one of its subclasses.

x

For the density function, x is a numeric vector of quantiles. For the moments functions, x is an object of class Distribution or one of its subclasses. For the log-likelihood and the estimation functions, x is the sample of observations.

...

extra arguments.

q

numeric. Vector of quantiles.

p

numeric. Vector of probabilities.

n

number of observations. If length(n) > 1, the length is taken to be the number required.

Value

Each type of function returns a different type of object:

Distribution Functions: When supplied with one argument (distr), the d(), p(), q(), r(), ll() functions return the density, cumulative probability, quantile, random sample generator, and log-likelihood functions, respectively. When supplied with both arguments (distr and x), they evaluate the aforementioned functions directly.

Moments: Returns a numeric, either vector or matrix depending on the moment and the distribution. The moments() function returns a list with all the available methods.

Estimation: Returns a list, the estimators of the unknown parameters. Note that in distribution families like the binomial, multinomial, and negative binomial, the size is not returned, since it is considered known.

Variance: Returns a named matrix. The asymptotic covariance matrix of the estimator.

Details

These S4 generic methods can work both as functions and as functionals (functions that return functions). The available distribution families are coded as S4 classes, specifically subclasses of the Distribution superclass. The methods can be used in two ways:

Option 1: If both the distr argument and x or n are supplied, then the function is evaluated directly, as usual.

Option 2: If only the distr argument is supplied, the method returns a function that takes as input the missing argument x or n, allowing the user to work with the function object itself. See examples.

Looking for a specific distribution family? This help page is general. Use the help page of each distribution to see the available methods for the class, details, and examples. Check the See Also section.

Functions

d(): density function

p(): cumulative distribution function

qn(): generalized inverse distribution function

r(): random sample generator function

See also

Examples

# -----------------------------------------------------
# Beta Distribution Example
# -----------------------------------------------------

# Create the distribution
a <- 3 ; b <- 5
D <- Beta(a, b)

# ------------------
# dpqr Functions
# ------------------

d(D, c(0.3, 0.8, 0.5)) # density function
#> [1] 2.268945 0.107520 1.640625
p(D, c(0.3, 0.8, 0.5)) # distribution function
#> [1] 0.3529305 0.9953280 0.7734375
qn(D, c(0.4, 0.8)) # inverse distribution function
#> [1] 0.3205858 0.5167578
x <- r(D, 100) # random generator function

# alternative way to use the function
df <- d(D) ; df(x) # df is a function itself
#>   [1] 2.2688239 2.0822803 1.5596763 1.2658636 2.1298589 1.9721568 2.2678160
#>   [8] 0.6188839 0.7372715 1.4928304 2.1499268 0.9990784 1.7734311 1.8801695
#>  [15] 2.0560081 2.1490984 1.9885426 2.2750859 2.2233764 0.2011700 2.2773168
#>  [22] 1.8810714 0.4837951 0.3835223 1.1104447 2.2154812 1.8131558 1.8588335
#>  [29] 1.4949584 2.2981017 1.9670601 0.9837032 2.0391066 2.2264502 1.7583458
#>  [36] 0.1538213 1.0138898 1.7693833 2.1339819 1.1298840 2.2580467 0.2945460
#>  [43] 2.2993173 1.5926982 2.0195034 2.0213303 2.1338020 1.8982922 2.0594952
#>  [50] 2.2917273 2.3022275 1.9902124 2.3015053 2.1832932 2.2279826 1.1646770
#>  [57] 1.8333530 1.5177388 1.3791370 2.1763079 1.8659916 1.1347729 1.7959103
#>  [64] 2.2836829 1.4367079 2.0781145 2.2777611 0.9543203 2.3032960 1.8661748
#>  [71] 1.6050071 1.3730708 0.7714459 0.6156657 1.7956211 2.2975234 2.2924535
#>  [78] 2.2154532 2.1185343 0.9609232 1.7518622 1.6764195 2.0484906 2.0391902
#>  [85] 1.9682763 2.2897504 1.4383587 0.5083482 2.2480721 2.1269161 0.9895356
#>  [92] 2.1883626 2.3014570 1.9681308 1.1502719 1.7561375 2.2357168 1.8403537
#>  [99] 2.2875825 2.2204210

# ------------------
# Moments
# ------------------

mean(D) # Expectation
#> [1] 0.375
var(D) # Variance
#> [1] 0.02604167
sd(D) # Standard Deviation
#> [1] 0.1613743
skew(D) # Skewness
#> [1] 0.3098387
kurt(D) # Excess Kurtosis
#> [1] 0.04
entro(D) # Entropy
#> [1] -0.4301508
finf(D) # Fisher Information Matrix
#>            shape1      shape2
#> shape1  0.2617971 -0.13313701
#> shape2 -0.1331370  0.08818594

# List of all available moments
mom <- moments(D)
mom$mean # expectation
#> [1] 0.375

# ------------------
# Point Estimation
# ------------------

ll(D, x)
#> [1] 45.75132
llbeta(x, a, b)
#> [1] 45.75132

ebeta(x, type = "mle")
#> $shape1
#> [1] 3.40893
#> 
#> $shape2
#> [1] 5.422875
#> 
ebeta(x, type = "me")
#> $shape1
#> [1] 3.427886
#> 
#> $shape2
#> [1] 5.444066
#> 
ebeta(x, type = "same")
#> $shape1
#> [1] 3.417856
#> 
#> $shape2
#> [1] 5.428137
#> 

mle(D, x)
#> $shape1
#> [1] 3.40893
#> 
#> $shape2
#> [1] 5.422875
#> 
me(D, x)
#> $shape1
#> [1] 3.427886
#> 
#> $shape2
#> [1] 5.444066
#> 
same(D, x)
#> $shape1
#> [1] 3.417856
#> 
#> $shape2
#> [1] 5.428137
#> 
e(D, x, type = "mle")
#> $shape1
#> [1] 3.40893
#> 
#> $shape2
#> [1] 5.422875
#> 

mle("beta", x) # the distr argument can be a character
#> $shape1
#> [1] 3.40893
#> 
#> $shape2
#> [1] 5.422875
#> 

# ------------------
# Estimator Variance
# ------------------

vbeta(a, b, type = "mle")
#>          shape1   shape2
#> shape1 16.44844 24.83272
#> shape2 24.83272 48.83039
vbeta(a, b, type = "me")
#>          shape1   shape2
#> shape1 17.64848 26.56970
#> shape2 26.56970 51.39394
vbeta(a, b, type = "same")
#>          shape1   shape2
#> shape1 16.57719 24.96198
#> shape2 24.96198 49.01071

avar_mle(D)
#>          shape1   shape2
#> shape1 16.44844 24.83272
#> shape2 24.83272 48.83039
avar_me(D)
#>          shape1   shape2
#> shape1 17.64848 26.56970
#> shape2 26.56970 51.39394
avar_same(D)
#>          shape1   shape2
#> shape1 16.57719 24.96198
#> shape2 24.96198 49.01071

v(D, type = "mle")
#>          shape1   shape2
#> shape1 16.44844 24.83272
#> shape2 24.83272 48.83039