Distribution Handling Helpers

This set of functions help handle the distribution classes. See Details.

Usage

get_moment_methods(x)

get_class_abbr(distr)

get_distr_class(distr)

s4_to_list(distr)

get_params(distr, list = TRUE)

get_unknown_params(distr, list = TRUE)

update_params(distr, prm, i)

Arguments

x, distr

an object of class Distribution.

list

logical. Should a list be returned? If FALSE, the return object is wrapped in unlist().

prm, i

A list containing three elements (name, pos, val) and the i-th element of val to be updated as the new parameter. See small_metrics().

Value

Depends on the function. See Details.

Functions

get_moment_methods(): Returns a character vector with the available moment methods for the distribution.

get_class_abbr(): Turns the S4 class in the name (character) used in the usual stats dpqr syntax.

get_distr_class(): Turns the distribution name from a character to an S4 class.

s4_to_list(): Turns an S4 distr object to a list.

get_params(): Get the parameters of a distribution as a list.

get_unknown_params(): Get the unknown parameters of a distribution as a list.

update_params(): Update the distribution parameters. Returns the distribution object. Used inside the small_metrics() and large_metrics() functions.

See also

Examples

if (FALSE) { # \dontrun{
distr <- Beta(1, 2)
get_moment_methods(distr)
get_class_abbr(distr)
get_distr_class("Gamma")
s4_to_list(distr)
get_params(distr, list = TRUE)
get_params(distr, list = FALSE)
get_unknown_params(distr)
} # }