Is it?

A set of functions that check whether an object has the desired characteristics.

Usage

is_whole(x, tol = .Machine$double.eps^0.5)

is_atvec(x)

is_numatvec(x)

is_nummat(x)

is_numarray(x)

is_symmetric(x)

is_pd(x)

is_pos(x)

is_integer(x)

is_natural(x)

Arguments

x

numeric vector or matrix.

tol

numeric. The tolerance for a numeric to be considered a whole number.

Value

logical. TRUE or FALSE, depending on whether the object satisfies the checks that define the characteristic.

Functions

is_whole(): Is the object an integer in the mathematical sense?

is_atvec(): Is the object an atomic vector (not matrix or array)?

is_numatvec(): Is the object a numeric atomic vector?

is_nummat(): Is the object a numeric atomic matrix?

is_numarray(): Is the object a numeric atomic array?

is_symmetric(): Is the object a symmetric matrix?

is_pd(): Is the object a positive definite matrix?

is_pos(): Are all the elements finite and positive?

is_integer(): Is the object an integer in the mathematical sense?

is_natural(): Is the object a natural number in the mathematical sense?

Examples

if (FALSE) { # \dontrun{
x<- matrix(1:9, nrow = 3, ncol = 3)
is_symmetric(x)
is_pd(x)
is_pos(x)
is_integer(x)
is_natural(x)
} # }