joker: Probability Distributions and Parameter Estimation

Implements an S4 distribution system and estimation methods for parameters of common distribution families. The common d, p, q, r function family for each distribution is enriched with the ll, e, and v counterparts, computing the log-likelihood, performing estimation, and calculating the asymptotic variance - covariance matrix, respectively. Parameter estimation is performed analytically whenever possible.

See also

Author

Maintainer: Ioannis Oikonomidis goikon@math.uoa.gr (ORCID)

Authors:

Samis Trevezas strevezas@math.uoa.gr (ORCID) [thesis advisor]