Progress Bar

Create a progress bar to be used with for loops that can possibly take a lot of time.

Usage

progress_bar(iter, total, start, message = NULL, width = 20)

format_hms(seconds)

Arguments

iter

integer. The current iteration step of the for loop.

total

integer. The total number of iterations.

start

POSIXct. The start time, as returned by Sys.time().

message

character. A message appearing before the progress bar.

width

integer. The length of the progress bar.

seconds

integer. Seconds to be converted into hh:mm:ss format.

Value

format_hms() returns a character in the hh:mm:ss format. progress_bar() prints the progress bar on the console, calling cat(), therefore as returns an invisible NULL.

Examples

if (FALSE) { # \dontrun{
# Turn seconds into hh:mm:ss
format_hms(7543)

n <- 100
start <- Sys.time()

for (i in 1:n) {
  Sys.sleep(0.1)  # Simulate time-consuming task
  progress_bar(i, n, start, message = "Computing", width = 30)
}
} # }